Initial commit1

+++ src/main/java/ru/kata/spring/boot_security/demo/Service/UserServiceImpl.java
@@ -2,8 +2,6 @@ package ru.kata.spring.boot_security.demo.Service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +9,7 @@ import ru.kata.spring.boot_security.demo.Repository.UserRepository;
 import ru.kata.spring.boot_security.demo.model.User;
 
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,38 +34,52 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void addUser(User user) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        userRepository.save(user);
+        if (userRepository.findByEmail(user.getUsername()).isEmpty()) {
+            User updatedUser = new User();
+            updatedUser.setPassword(passwordEncoder.encode(user.getPassword()));
+            updatedUser.setAge(user.getAge());
+            updatedUser.setId(user.getId());
+            updatedUser.setName(user.getName());
+            updatedUser.setLastName(user.getLastName());
+            updatedUser.setEmail(user.getEmail());
+            updatedUser.setRoles(user.getRoles());
+            userRepository.save(updatedUser);
+        } else {
+            throw new EntityNotFoundException("не получилось добавить");
+        }
     }
 
     @Override
     public User userID(long id) {
         Optional<User> optional = userRepository.findById(id);
-        return optional.orElseThrow();
+        return optional.orElseThrow(() -> new EntityNotFoundException("не получилось получить, получается"));
     }
 
     @Transactional
     @Override
     public void removeUserById(long id) {
-        userRepository.deleteById(id);
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("не получилось удалить");
+        }
     }
 
     @Transactional
     @Override
     public void update(User user, long id) {
-        User updatedUser = userRepository.findById(id).get();
-        updatedUser.setName(user.getName());
-        updatedUser.setRoles(user.getRoles());
-        updatedUser.setPassword(passwordEncoder.encode(user.getPassword()));
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<User> updatedUser = userRepository.findByName(username);
-        if (updatedUser.isEmpty()) {
-            throw new UsernameNotFoundException("Нет такого");
+        if (userRepository.findById(id).isPresent()) {
+            User updatedUser = userRepository.findById(id).get();
+            updatedUser.setName(user.getName());
+            updatedUser.setRoles(user.getRoles());
+            updatedUser.setEmail(user.getEmail());
+            updatedUser.setLastName(user.getLastName());
+            updatedUser.setAge(user.getAge());
+            if (!user.getPassword().isEmpty()) {
+                updatedUser.setPassword(passwordEncoder.encode(user.getPassword()));
+            }
+        } else {
+            throw new EntityNotFoundException("не получилось обновить");
         }
-        return updatedUser.get();
     }
-
 }
